fix: декомпозировал функции в useeffect
